Refactoring Phase 1: split oversized files into focused modules
1.1 Split datasets/loader.py (1766 lines):
- Extract RestrictedUnpickler + safe_pickle_load → datasets/pickle_security.py
- Extract 8 metadata functions → datasets/metadata.py
- Move _apply_converter_to_demo → converters/runner.py
- loader.py re-exports for backward compatibility

1.2 Split routes/sorting.py (1138 lines) into focused blueprints:
- Label endpoints → routes/labels.py (labels_bp)
- Server media files → routes/media_server.py (media_server_bp)
- Evaluation endpoints → routes/eval.py (eval_bp)
- Core sort/voting remains in sorting.py

1.3 Extract business logic from routes/trainable_models.py (1194 lines):
- _seed_good_votes_from_examples → models/media_seeding.py
- _restore_labels_from_trainable_model → models/label_restoration.py
- _apply_and_retrain → models/training_workflow.py
- trainable_models.py imports from new locations (backward compat)
