+++ vtsearch/converters/runner.py
@@ -268,3 +268,124 @@ def _compute_md5(output: dict[str, Any]) -> str:
     if text:
         return hashlib.md5(text.encode("utf-8")).hexdigest()
     return hashlib.md5(b"").hexdigest()
+
+
+def apply_converter_to_demo(
+    converter_name: str,
+    dataset_name: str,
+    medias: dict[int, dict[str, Any]],
+    embedder_name: str = "",
+    on_progress: Optional[ProgressCallback] = None,
+) -> None:
+    """Convert all medias in-place using the named converter.
+
+    After conversion, *medias* contains the converted outputs (target type)
+    instead of the original source-type medias.  Each converted media's
+    origin records the demo dataset and the converter used.
+    """
+    converter = get_converter(converter_name)
+    if converter is None:
+        raise ValueError(f"Unknown converter: {converter_name}")
+
+    if on_progress is None:
+        on_progress = _default_progress()
+
+    from vtsearch.media import embedders_for_type, get_embedder  # noqa: PLC0415
+
+    # Resolve embedder for the *target* type.
+    target_emb = None
+    if embedder_name:
+        try:
+            target_emb = get_embedder(embedder_name)
+        except KeyError:
+            pass
+    if target_emb is None:
+        avail = embedders_for_type(converter.target_type)
+        if avail:
+            target_emb = avail[0]
+    if target_emb is not None and getattr(target_emb, "_model", None) is None:
+        target_emb.load_models()
+
+    # Convert each original media and collect the outputs.
+    source_items = list(medias.items())
+    converted: dict[int, dict[str, Any]] = {}
+    new_id = 1
+
+    on_progress(
+        "converting",
+        f"Converting {len(source_items)} items via {converter.display_name or converter.name}...",
+        0,
+        len(source_items),
+    )
+
+    for idx, (_, src) in enumerate(source_items):
+        on_progress(
+            "converting",
+            f"Converting {src.get('filename', '')} via {converter.display_name or converter.name}...",
+            idx + 1,
+            len(source_items),
+        )
+
+        try:
+            outputs = converter.convert(src)
+        except Exception:
+            continue
+        if not outputs:
+            continue
+
+        origin = {
+            "importer": "converter",
+            "params": {
+                "converter": converter_name,
+                "source_file": src.get("filename", ""),
+                "parent_importer": "demo",
+                "parent_demo": dataset_name,
+            },
+        }
+
+        for output in outputs:
+            output_filename = output.get("filename", f"converted_{new_id}")
+            source_name = src.get("filename", str(src.get("id", "")))
+            origin_name = f"{source_name}\u2192{output_filename}"
+
+            embedding = _embed_converted_output(target_emb, output)
+            if embedding is None:
+                continue
+
+            md5 = _compute_md5(output)
+
+            media_data: dict[str, Any] = {
+                "id": new_id,
+                "type": converter.target_type,
+                "embedder": target_emb.name if target_emb else "",
+                "file_size": len(output.get("media_bytes", b"") or output.get("media_string", "").encode()),
+                "md5": md5,
+                "embedding": embedding,
+                "filename": origin_name,
+                "category": src.get("category", "custom"),
+                "origin": origin,
+                "origin_name": origin_name,
+                "media_bytes": None,
+                "media_string": None,
+                "media_path": src.get("media_path", ""),
+                "duration": output.get("duration", 0),
+            }
+
+            if "media_bytes" in output:
+                media_data["media_bytes"] = output["media_bytes"]
+            if "media_string" in output:
+                media_data["media_string"] = output["media_string"]
+            if "width" in output:
+                media_data["width"] = output["width"]
+            if "height" in output:
+                media_data["height"] = output["height"]
+            if "word_count" in output:
+                media_data["word_count"] = output["word_count"]
+            if "character_count" in output:
+                media_data["character_count"] = output["character_count"]
+
+            converted[new_id] = media_data
+            new_id += 1
+
+    medias.clear()
+    medias.update(converted)
