+++ vtsearch/routes/__init__.py
@@ -1,7 +1,10 @@
 """Flask blueprints for organizing application routes."""
 
 from vtsearch.routes.auth import auth_bp
+from vtsearch.routes.eval import eval_bp
 from vtsearch.routes.file_browser import file_browser_bp
+from vtsearch.routes.labels import labels_bp
+from vtsearch.routes.media_server import media_server_bp
 from vtsearch.routes.medias import medias_bp
 from vtsearch.routes.datasets import datasets_bp
 from vtsearch.routes.detectors import detectors_bp
@@ -15,7 +18,10 @@ from vtsearch.routes.trainable_models import trainable_models_bp
 
 __all__ = [
     "auth_bp",
+    "eval_bp",
     "file_browser_bp",
+    "labels_bp",
+    "media_server_bp",
     "main_bp",
     "medias_bp",
     "sorting_bp",
